requirements.txt
alembic==1.7.5
click==8.0.1
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.6
MarkupSafe==2.0.1
mysqlclient==2.0.3
PyMySQL==1.0.2
PyYAML==6.0
SQLAlchemy==1.4.31
Werkzeug==2.0.1
